# Connect-Hub/lib/api-zod/src/generated/types/analyticsOverview.zod.ts
/**
 * Connect-Hub orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MeetNow API specification
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const AnalyticsOverview = zod.object({
  meetingsThisMonth: zod.number(),
  totalHours: zod.string(),
  peopleMet: zod.number(),
  messagesSent: zod.number(),
});

export type AnalyticsOverview = zod.input<typeof AnalyticsOverview>;